# app/endpoints/openapi.yaml
...
openapi: 3.0.3
info:
  title: REST API
  version: v1.0.0
paths:
...
  /vrf/v1/info:
    get:
      description: |-
        Info returns static drand chain information used for verification and
         debugging.
      operationId: v1_info
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryInfoResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Info
      tags:
        - V1
...
  /vrf/v1/randomness:
    get:
      description: |-
        Randomness returns a drand beacon for a specific round or the latest
         available beacon when round is unset (zero).
      operationId: v1_randomness
      parameters:
        - name: round
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRandomnessResponse'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
      summary: Randomness
      tags:
        - V1
components:
  schemas:
...
    GoogleProtobufAny:
      type: object
      properties:
        '@type':
          type: string
          description: The type of the serialized message.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
    QueryInfoResponse:
      type: object
      properties:
        chainHash:
          type: string
          description: chain_hash is the canonical identifier for the drand chain.
          format: bytes
        publicKey:
          type: string
          description: public_key is the collective BLS public key for the drand group.
          format: bytes
        periodSeconds:
          type: string
          description: period_seconds is the drand beacon period in seconds.
        genesisUnixSec:
          type: string
          description: genesis_unix_sec is the drand genesis time (UNIX seconds) for round 1.
      description: QueryInfoResponse carries static drand chain information.
...
    QueryRandomnessResponse:
      type: object
      properties:
        drandRound:
          type: string
          description: drand_round is the drand round number.
        randomness:
          type: string
          description: randomness is SHA256(signature).
          format: bytes
        signature:
          type: string
          description: signature is the drand BLS signature for this round.
          format: bytes
        previousSignature:
          type: string
          description: |-
            previous_signature is the BLS signature of the previous round, required
             for chained-scheme verification.
          format: bytes
      description: QueryRandomnessResponse carries a single drand beacon.
...
    Status:
      type: object
      properties:
        code:
          type: integer
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
          format: int32
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/GoogleProtobufAny'
          description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
      description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...